api/pyproject.toml
[tool.poetry]
name = "base"
version = "0.1.0"
description = ""
authors = ["tbson"]

[tool.poetry.dependencies]
Django = "^4.0.4"
django-filter = "^21.1"
django-phonenumber-field = "^6.1.0"
django-rest-auth = "^0.9.5"
djangorestframework = "^3.13.1"
djangorestframework-simplejwt = "^5.1.0"
phonenumbers = "^8.12.47"
Pillow = "^9.1.0"
psycopg2-binary = "^2.9.3"
python = "^3.10"
requests = "^2.27.1"
argon2-cffi = "^21.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
